Trait mc_full_service::db::account::AccountModel

pub trait AccountModel {

    fn create_from_mnemonic(
        mnemonic: &Mnemonic,
        first_block_index: Option<u64>,
        import_block_index: Option<u64>,
        next_subaddress_index: Option<u64>,
        name: &str,
        fog_report_url: String,
        fog_report_id: String,
        fog_authority_spki: String,
        conn: &Conn
    ) -> Result<(AccountID, String), WalletDbError>;
    fn create_from_root_entropy(
        entropy: &RootEntropy,
        first_block_index: Option<u64>,
        import_block_index: Option<u64>,
        next_subaddress_index: Option<u64>,
        name: &str,
        fog_report_url: String,
        fog_report_id: String,
        fog_authority_spki: String,
        conn: &Conn
    ) -> Result<(AccountID, String), WalletDbError>;
    fn create(
        entropy: &[u8],
        key_derivation_version: u8,
        account_key: &AccountKey,
        first_block_index: Option<u64>,
        import_block_index: Option<u64>,
        next_subaddress_index: Option<u64>,
        name: &str,
        fog_enabled: bool,
        conn: &Conn
    ) -> Result<(AccountID, String), WalletDbError>;
    fn import(
        mnemonic: &Mnemonic,
        name: Option<String>,
        import_block_index: u64,
        first_block_index: Option<u64>,
        next_subaddress_index: Option<u64>,
        fog_report_url: String,
        fog_report_id: String,
        fog_authority_spki: String,
        conn: &Conn
    ) -> Result<Account, WalletDbError>;
    fn import_legacy(
        entropy: &RootEntropy,
        name: Option<String>,
        import_block_index: u64,
        first_block_index: Option<u64>,
        next_subaddress_index: Option<u64>,
        fog_report_url: String,
        fog_report_id: String,
        fog_authority_spki: String,
        conn: &Conn
    ) -> Result<Account, WalletDbError>;
    fn import_view_only(
        view_private_key: &RistrettoPrivate,
        spend_public_key: &RistrettoPublic,
        name: Option<String>,
        import_block_index: u64,
        first_block_index: Option<u64>,
        next_subaddress_index: Option<u64>,
        conn: &Conn
    ) -> Result<Account, WalletDbError>;
    fn list_all(
        conn: &Conn,
        offset: Option<u64>,
        limit: Option<u64>
    ) -> Result<Vec<Account>, WalletDbError>;
    fn get(account_id: &AccountID, conn: &Conn) -> Result<Account, WalletDbError>;
    fn get_by_txo_id(
        txo_id_hex: &str,
        conn: &Conn
    ) -> Result<Vec<Account>, WalletDbError>;
    fn update_name(
        &self,
        new_name: String,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn update_next_block_index(
        &self,
        next_block_index: u64,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn delete(self, conn: &Conn) -> Result<(), WalletDbError>;
    fn change_subaddress(
        self,
        conn: &Conn
    ) -> Result<AssignedSubaddress, WalletDbError>;
    fn main_subaddress(
        self,
        conn: &Conn
    ) -> Result<AssignedSubaddress, WalletDbError>;
    fn get_token_ids(self, conn: &Conn) -> Result<Vec<TokenId>, WalletDbError>;
    fn next_subaddress_index(self, conn: &Conn) -> Result<u64, WalletDbError>;
    fn account_key(&self) -> Result<Option<AccountKey>, WalletDbError>;
    fn view_account_key(&self) -> Result<ViewAccountKey, WalletDbError>;
    fn view_private_key(&self) -> Result<RistrettoPrivate, WalletDbError>;

}

Required Methods

fn create_from_mnemonic(
mnemonic: &Mnemonic,
first_block_index: Option<u64>,
import_block_index: Option<u64>,
next_subaddress_index: Option<u64>,
name: &str,
fog_report_url: String,
fog_report_id: String,
fog_authority_spki: String,
conn: &Conn
) -> Result<(AccountID, String), WalletDbError>

Create an account.

Returns:

(account_id, main_subaddress_b58)

fn create_from_root_entropy(
entropy: &RootEntropy,
first_block_index: Option<u64>,
import_block_index: Option<u64>,
next_subaddress_index: Option<u64>,
name: &str,
fog_report_url: String,
fog_report_id: String,
fog_authority_spki: String,
conn: &Conn
) -> Result<(AccountID, String), WalletDbError>

Create an account.

Returns:

(account_id, main_subaddress_b58)

fn create(
entropy: &[u8],
key_derivation_version: u8,
account_key: &AccountKey,
first_block_index: Option<u64>,
import_block_index: Option<u64>,
next_subaddress_index: Option<u64>,
name: &str,
fog_enabled: bool,
conn: &Conn
) -> Result<(AccountID, String), WalletDbError>

Create an account.

Returns:

(account_id, main_subaddress_b58)

fn import(
mnemonic: &Mnemonic,
name: Option<String>,
import_block_index: u64,
first_block_index: Option<u64>,
next_subaddress_index: Option<u64>,
fog_report_url: String,
fog_report_id: String,
fog_authority_spki: String,
conn: &Conn
) -> Result<Account, WalletDbError>

Import account.

fn import_legacy(
entropy: &RootEntropy,
name: Option<String>,
import_block_index: u64,
first_block_index: Option<u64>,
next_subaddress_index: Option<u64>,
fog_report_url: String,
fog_report_id: String,
fog_authority_spki: String,
conn: &Conn
) -> Result<Account, WalletDbError>

Import account.

fn import_view_only(
view_private_key: &RistrettoPrivate,
spend_public_key: &RistrettoPublic,
name: Option<String>,
import_block_index: u64,
first_block_index: Option<u64>,
next_subaddress_index: Option<u64>,
conn: &Conn
) -> Result<Account, WalletDbError>

Import a view only account.

fn list_all(
conn: &Conn,
offset: Option<u64>,
limit: Option<u64>
) -> Result<Vec<Account>, WalletDbError>

List all accounts.

Returns:

Vector of all Accounts in the DB

fn get(account_id: &AccountID, conn: &Conn) -> Result<Account, WalletDbError>

Get a specific account.

Returns:

Account

fn get_by_txo_id(
txo_id_hex: &str,
conn: &Conn
) -> Result<Vec<Account>, WalletDbError>

Get the accounts associated with the given Txo.

fn update_name(&self, new_name: String, conn: &Conn) -> Result<(), WalletDbError>

Update an account. The only updatable field is the name. Any other desired update requires adding a new account, and deleting the existing if desired.

fn update_next_block_index(
&self,
next_block_index: u64,
conn: &Conn
) -> Result<(), WalletDbError>

Update the next block index this account will need to sync.

fn delete(self, conn: &Conn) -> Result<(), WalletDbError>

Delete an account.

fn change_subaddress(
self,
conn: &Conn
) -> Result<AssignedSubaddress, WalletDbError>

Get change public address

fn main_subaddress(
self,
conn: &Conn
) -> Result<AssignedSubaddress, WalletDbError>

Get main public address

fn get_token_ids(self, conn: &Conn) -> Result<Vec<TokenId>, WalletDbError>

Get all of the token ids present for the account

fn next_subaddress_index(self, conn: &Conn) -> Result<u64, WalletDbError>

Get the next sequentially unassigned subaddress index for the account (reserved addresses are not included)

fn account_key(&self) -> Result<Option<AccountKey>, WalletDbError>

fn view_account_key(&self) -> Result<ViewAccountKey, WalletDbError>

fn view_private_key(&self) -> Result<RistrettoPrivate, WalletDbError>