Trait mc_full_service::db::assigned_subaddress::AssignedSubaddressModel

pub trait AssignedSubaddressModel {
    fn create(
        account_key: &AccountKey,
        subaddress_index: u64,
        comment: &str,
        conn: &Conn
    ) -> Result<String, WalletDbError>;
    fn create_for_view_only_account(
        account_key: &ViewAccountKey,
        subaddress_index: u64,
        comment: &str,
        conn: &Conn
    ) -> Result<String, WalletDbError>;
    fn create_next_for_account(
        account_id_hex: &str,
        comment: &str,
        ledger_db: &LedgerDB,
        conn: &Conn
    ) -> Result<(String, i64), WalletDbError>;
    fn get(
        public_address_b58: &str,
        conn: &Conn
    ) -> Result<AssignedSubaddress, WalletDbError>;
    fn get_for_account_by_index(
        account_id_hex: &str,
        index: i64,
        conn: &Conn
    ) -> Result<AssignedSubaddress, WalletDbError>;
    fn find_by_subaddress_spend_public_key(
        subaddress_spend_public_key: &RistrettoPublic,
        conn: &Conn
    ) -> Result<(i64, String), WalletDbError>;
    fn list_all(
        account_id: Option<String>,
        offset: Option<u64>,
        limit: Option<u64>,
        conn: &Conn
    ) -> Result<Vec<AssignedSubaddress>, WalletDbError>;
    fn delete_all(account_id_hex: &str, conn: &Conn) -> Result<(), WalletDbError>;
    fn public_address(self) -> Result<PublicAddress, WalletDbError>;
}

Required Methods

fn create(
account_key: &AccountKey,
subaddress_index: u64,
comment: &str,
conn: &Conn
) -> Result<String, WalletDbError>

Assign a subaddress to a contact.

Inserts an AssignedSubaddress to the DB.

Arguments

account_key - An account’s private keys.

address_book_entry -

subaddress_index -

comment -

conn -

Returns

public_address_b58

fn create_for_view_only_account(
account_key: &ViewAccountKey,
subaddress_index: u64,
comment: &str,
conn: &Conn
) -> Result<String, WalletDbError>

fn create_next_for_account(
account_id_hex: &str,
comment: &str,
ledger_db: &LedgerDB,
conn: &Conn
) -> Result<(String, i64), WalletDbError>

Create the next subaddress for a given account.

Returns:

(public_address_b58, subaddress_index)

fn get(
public_address_b58: &str,
conn: &Conn
) -> Result<AssignedSubaddress, WalletDbError>

Get the AssignedSubaddress for a given public_address_b58

fn get_for_account_by_index(
account_id_hex: &str,
index: i64,
conn: &Conn
) -> Result<AssignedSubaddress, WalletDbError>

Get the Assigned Subaddress for a given index in an account, if it exists

fn find_by_subaddress_spend_public_key(
subaddress_spend_public_key: &RistrettoPublic,
conn: &Conn
) -> Result<(i64, String), WalletDbError>

Find an AssignedSubaddress by the subaddress spend public key

Returns:

(subaddress_index, public_address_b58)

fn list_all(
account_id: Option<String>,
offset: Option<u64>,
limit: Option<u64>,
conn: &Conn
) -> Result<Vec<AssignedSubaddress>, WalletDbError>

List all AssignedSubaddresses for a given account.

fn delete_all(account_id_hex: &str, conn: &Conn) -> Result<(), WalletDbError>

Delete all AssignedSubaddresses for a given account.

fn public_address(self) -> Result<PublicAddress, WalletDbError>

Helper to get the public address out of the assigned subaddress