Trait mc_full_service::db::gift_code::GiftCodeModel

pub trait GiftCodeModel {
    fn create(
        gift_code_b58: &EncodedGiftCode,
        value: i64,
        conn: &Conn
    ) -> Result<GiftCode, WalletDbError>;
    fn get(
        gift_code_b58: &EncodedGiftCode,
        conn: &Conn
    ) -> Result<GiftCode, WalletDbError>;
    fn list_all(
        conn: &Conn,
        offset: Option<u64>,
        limit: Option<u64>
    ) -> Result<Vec<GiftCode>, WalletDbError>;
    fn delete(self, conn: &Conn) -> Result<(), WalletDbError>;
}

Required Methods

fn create(
gift_code_b58: &EncodedGiftCode,
value: i64,
conn: &Conn
) -> Result<GiftCode, WalletDbError>

Create a gift code.

Gift code includes:

entropy

txo public key

memo

GiftCodeModel::create should be called after the account has already been inserted into the DB, the txo has already been deposited to that account, and the transaction_log has been stored for that deposit, all of which are handled by the GiftCodeService.

Returns:

Gift code encoded as b58 string.

fn get(
gift_code_b58: &EncodedGiftCode,
conn: &Conn
) -> Result<GiftCode, WalletDbError>

Get the details of a specific Gift Code.

fn list_all(
conn: &Conn,
offset: Option<u64>,
limit: Option<u64>
) -> Result<Vec<GiftCode>, WalletDbError>

Get all Gift Codes in this wallet.

fn delete(self, conn: &Conn) -> Result<(), WalletDbError>

Delete a gift code.