Module mc_full_service::db::models

DB Models

Structs

Account

An Account entity.

AssignedSubaddress

A subaddress given to a particular contact, for the purpose of tracking funds received from that contact.

GiftCode

Migration

NewAccount

A structure that can be inserted to create a new entity in the accounts table.

NewAssignedSubaddress

A structure that can be inserted to create a new AssignedSubaddress entity.

NewGiftCode

NewMigration

NewTransactionInputTxo

NewTransactionLog

A structure that can be inserted to create a new TransactionLog entity.

NewTransactionOutputTxo

NewTxo

A structure that can be inserted to create a new entity in the txos table.

TransactionInputTxo

TransactionLog

The status of a sent transaction OR a received transaction output.

TransactionOutputTxo

Txo

A transaction output entity that either was received to an Account in this wallet, or originated from an Account in this wallet. A transaction output can be in one of many states with respect to multiple accounts. Managing these relationships and states is one of the main goals of the Full-Service wallet.