Struct mc_full_service::db::models::Txo

pub struct Txo {
    pub id: String,
    pub account_id: Option<String>,
    pub value: i64,
    pub token_id: i64,
    pub target_key: Vec<u8>,
    pub public_key: Vec<u8>,
    pub e_fog_hint: Vec<u8>,
    pub subaddress_index: Option<i64>,
    pub key_image: Option<Vec<u8>>,
    pub received_block_index: Option<i64>,
    pub spent_block_index: Option<i64>,
    pub confirmation: Option<Vec<u8>>,
}

A transaction output entity that either was received to an Account in this wallet, or originated from an Account in this wallet. A transaction output can be in one of many states with respect to multiple accounts. Managing these relationships and states is one of the main goals of the Full-Service wallet.

Fields
id: String

Primary key derived from the contents of the ledger TxOut

account_id: Option<String>value: i64

The value of this transaction output, in picoMob.

token_id: i64

The token of this transaction output.

target_key: Vec<u8>

The serialized target_key of the TxOut.

public_key: Vec<u8>

The serialized public_key of the TxOut.

e_fog_hint: Vec<u8>

The serialized e_fog_hint of the TxOut.

subaddress_index: Option<i64>

The receiving subaddress, if known.

key_image: Option<Vec<u8>>

Pre-computed key image for this Txo, or None if the Txo is orphaned.

received_block_index: Option<i64>

Block index containing this Txo.

spent_block_index: Option<i64>confirmation: Option<Vec<u8>>
Implementations

impl Txo

pub fn public_key(&self) -> Result<CompressedRistrettoPublic, DecodeError>

impl Txo

pub fn amount(&self) -> Amount