Trait mc_full_service::db::transaction_log::TransactionLogModel

pub trait TransactionLogModel {

    fn get(
        id: &TransactionId,
        conn: &Conn
    ) -> Result<TransactionLog, WalletDbError>;
    fn get_associated_txos(
        &self,
        conn: &Conn
    ) -> Result<AssociatedTxos, WalletDbError>;
    fn update_submitted_block_index(
        &self,
        submitted_block_index: u64,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn update_comment(
        &self,
        comment: String,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn update_tx_and_tombstone_block_index(
        &self,
        tx: &[u8],
        tombstone_block_index: Option<i64>,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn list_all(
        account_id: Option<String>,
        offset: Option<u64>,
        limit: Option<u64>,
        min_block_index: Option<u64>,
        max_block_index: Option<u64>,
        conn: &Conn
    ) -> Result<Vec<(TransactionLog, AssociatedTxos, ValueMap)>, WalletDbError>;
    fn log_built(
        unsigned_tx_proposal: &UnsignedTxProposal,
        account_id: &AccountID,
        conn: &Conn
    ) -> Result<TransactionLog, WalletDbError>;
    fn log_signed(
        tx_proposal: TxProposal,
        comment: String,
        account_id_hex: &str,
        conn: &Conn
    ) -> Result<TransactionLog, WalletDbError>;
    fn log_submitted(
        tx_proposal: &TxProposal,
        block_index: u64,
        comment: String,
        account_id_hex: &str,
        conn: &Conn
    ) -> Result<TransactionLog, WalletDbError>;
    fn delete_all_for_account(
        account_id_hex: &str,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn update_pending_associated_with_txo_to_succeeded(
        txo_id_hex: &str,
        finalized_block_index: u64,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn update_pending_exceeding_tombstone_block_index_to_failed(
        block_index: u64,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn status(&self) -> TxStatus;
    fn value_for_token_id(
        &self,
        token_id: TokenId,
        conn: &Conn
    ) -> Result<u64, WalletDbError>;
    fn value_map(&self, conn: &Conn) -> Result<ValueMap, WalletDbError>;

}

Required Methods

fn get(id: &TransactionId, conn: &Conn) -> Result<TransactionLog, WalletDbError>

Get a transaction log from the TransactionId.

fn get_associated_txos(
&self,
conn: &Conn
) -> Result<AssociatedTxos, WalletDbError>

Get the Txos associated with a given TransactionId, grouped according to their type.

Returns:

AssoiatedTxos(inputs, outputs, change)

fn update_submitted_block_index(
&self,
submitted_block_index: u64,
conn: &Conn
) -> Result<(), WalletDbError>

fn update_comment(
&self,
comment: String,
conn: &Conn
) -> Result<(), WalletDbError>

fn update_tx_and_tombstone_block_index(
&self,
tx: &[u8],
tombstone_block_index: Option<i64>,
conn: &Conn
) -> Result<(), WalletDbError>

fn list_all(
account_id: Option<String>,
offset: Option<u64>,
limit: Option<u64>,
min_block_index: Option<u64>,
max_block_index: Option<u64>,
conn: &Conn
) -> Result<Vec<(TransactionLog, AssociatedTxos, ValueMap)>, WalletDbError>

List all TransactionLogs and their associated Txos for a given account.

Returns:

Vec(TransactionLog, AssociatedTxos(inputs, outputs, change))

fn log_built(
unsigned_tx_proposal: &UnsignedTxProposal,
account_id: &AccountID,
conn: &Conn
) -> Result<TransactionLog, WalletDbError>

Log a transaction that has been built but not yet signed.

Returns:

TransactionLog

fn log_signed(
tx_proposal: TxProposal,
comment: String,
account_id_hex: &str,
conn: &Conn
) -> Result<TransactionLog, WalletDbError>

Log a transaction that has been signed

Returns:

TransactionLog

fn log_submitted(
tx_proposal: &TxProposal,
block_index: u64,
comment: String,
account_id_hex: &str,
conn: &Conn
) -> Result<TransactionLog, WalletDbError>

Log a submitted transaction.

When submitting a transaction, we store relevant information to the transaction logs, and we also track information about each of the txos involved in the transaction.

Note: We expect transactions created with this wallet to have one recipient, with the rest of the minted txos designated as change. Other wallets may choose to behave differently, but our TransactionLogs Table assumes this behavior.

fn delete_all_for_account(
account_id_hex: &str,
conn: &Conn
) -> Result<(), WalletDbError>

Remove all logs for an account

fn update_pending_associated_with_txo_to_succeeded(
txo_id_hex: &str,
finalized_block_index: u64,
conn: &Conn
) -> Result<(), WalletDbError>

fn update_pending_exceeding_tombstone_block_index_to_failed(
block_index: u64,
conn: &Conn
) -> Result<(), WalletDbError>

fn status(&self) -> TxStatus

fn value_for_token_id(
&self,
token_id: TokenId,
conn: &Conn
) -> Result<u64, WalletDbError>

fn value_map(&self, conn: &Conn) -> Result<ValueMap, WalletDbError>