Trait mc_full_service::db::txo::TxoModel

pub trait TxoModel {

    fn create_received(
        tx_out: TxOut,
        subaddress_index: Option<u64>,
        key_image: Option<KeyImage>,
        amount: Amount,
        received_block_index: u64,
        account_id_hex: &str,
        conn: &Conn
    ) -> Result<String, WalletDbError>;
    fn create_new_output(
        output_txo: &OutputTxo,
        is_change: bool,
        transaction_id: &TransactionId,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn update_as_received(
        &self,
        subaddress_index: Option<u64>,
        received_key_image: Option<KeyImage>,
        block_index: u64,
        account_id_hex: &str,
        amount: Amount,
        target_key: &Vec<u8>,
        public_key: &Vec<u8>,
        e_fog_hint: &Vec<u8>,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn update_spent_block_index(
        txo_id_hex: &str,
        spent_block_index: u64,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn update_key_image(
        txo_id_hex: &str,
        key_image: &KeyImage,
        spent_block_index: Option<u64>,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn list(
        status: Option<TxoStatus>,
        min_received_block_index: Option<u64>,
        max_received_block_index: Option<u64>,
        offset: Option<u64>,
        limit: Option<u64>,
        token_id: Option<u64>,
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn list_for_account(
        account_id_hex: &str,
        status: Option<TxoStatus>,
        min_received_block_index: Option<u64>,
        max_received_block_index: Option<u64>,
        offset: Option<u64>,
        limit: Option<u64>,
        token_id: Option<u64>,
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn list_for_address(
        assigned_subaddress_b58: &str,
        status: Option<TxoStatus>,
        min_received_block_index: Option<u64>,
        max_received_block_index: Option<u64>,
        offset: Option<u64>,
        limit: Option<u64>,
        token_id: Option<u64>,
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn list_unspent_or_pending_key_images(
        account_id_hex: &str,
        token_id: Option<u64>,
        conn: &Conn
    ) -> Result<HashMap<KeyImage, String>, WalletDbError>;
    fn list_unspent(
        account_id_hex: Option<&str>,
        assigned_subaddress_b58: Option<&str>,
        token_id: Option<u64>,
        min_received_block_index: Option<u64>,
        max_received_block_index: Option<u64>,
        offset: Option<u64>,
        limit: Option<u64>,
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn list_spent(
        account_id_hex: Option<&str>,
        assigned_subaddress_b58: Option<&str>,
        token_id: Option<u64>,
        min_received_block_index: Option<u64>,
        max_received_block_index: Option<u64>,
        offset: Option<u64>,
        limit: Option<u64>,
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn list_orphaned(
        account_id_hex: Option<&str>,
        token_id: Option<u64>,
        min_received_block_index: Option<u64>,
        max_received_block_index: Option<u64>,
        offset: Option<u64>,
        limit: Option<u64>,
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn list_pending(
        account_id_hex: Option<&str>,
        assigned_subaddress_b58: Option<&str>,
        token_id: Option<u64>,
        min_received_block_index: Option<u64>,
        max_received_block_index: Option<u64>,
        offset: Option<u64>,
        limit: Option<u64>,
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn list_unverified(
        account_id_hex: Option<&str>,
        assigned_subaddress_b58: Option<&str>,
        token_id: Option<u64>,
        min_received_block_index: Option<u64>,
        max_received_block_index: Option<u64>,
        offset: Option<u64>,
        limit: Option<u64>,
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn list_spendable(
        account_id_hex: Option<&str>,
        max_spendable_value: Option<u64>,
        assigned_subaddress_b58: Option<&str>,
        token_id: u64,
        default_token_fee: u64,
        conn: &Conn
    ) -> Result<SpendableTxosResult, WalletDbError>;
    fn list_created(
        account_id_hex: Option<&str>,
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn list_secreted(
        account_id_hex: Option<&str>,
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn get(txo_id_hex: &str, conn: &Conn) -> Result<Txo, WalletDbError>;
    fn select_by_public_key(
        public_keys: &[&CompressedRistrettoPublic],
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn select_by_id(
        txo_ids: &[String],
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn select_spendable_txos_for_value(
        account_id_hex: &str,
        target_value: u128,
        max_spendable_value: Option<u64>,
        token_id: u64,
        default_token_fee: u64,
        conn: &Conn
    ) -> Result<Vec<Txo>, WalletDbError>;
    fn validate_confirmation(
        account_id: &AccountID,
        txo_id_hex: &str,
        confirmation: &TxOutConfirmationNumber,
        conn: &Conn
    ) -> Result<bool, WalletDbError>;
    fn scrub_account(
        account_id_hex: &str,
        conn: &Conn
    ) -> Result<(), WalletDbError>;
    fn delete_unreferenced(conn: &Conn) -> Result<(), WalletDbError>;
    fn status(&self, conn: &Conn) -> Result<TxoStatus, WalletDbError>;
    fn membership_proof(
        &self,
        ledger_db: &LedgerDB
    ) -> Result<TxOutMembershipProof, WalletDbError>;

}

Required Methods

fn create_received(
tx_out: TxOut,
subaddress_index: Option<u64>,
key_image: Option<KeyImage>,
amount: Amount,
received_block_index: u64,
account_id_hex: &str,
conn: &Conn
) -> Result<String, WalletDbError>

Upserts a received Txo.

Arguments

txo - a TxOut contained in the ledger.

subaddress_index - The receiving subaddress index, if known.

key_image -

value - The value of the output, in picoMob.

received_block_index - the block at which the Txo was received.

account_id_hex - the account ID for the account which received this Txo.

conn - Sqlite database connection.

The subaddress_index may be None, and the Txo is said to be “orphaned”, if the subaddress is not yet being tracked by the wallet.

Returns

txo_id_hex

fn create_new_output(
output_txo: &OutputTxo,
is_change: bool,
transaction_id: &TransactionId,
conn: &Conn
) -> Result<(), WalletDbError>

fn update_as_received(
&self,
subaddress_index: Option<u64>,
received_key_image: Option<KeyImage>,
block_index: u64,
account_id_hex: &str,
amount: Amount,
target_key: &Vec<u8>,
public_key: &Vec<u8>,
e_fog_hint: &Vec<u8>,
conn: &Conn
) -> Result<(), WalletDbError>

Update an existing Txo to spendable by including its subaddress_index and optionally the key_image in the case of view only accounts.

fn update_spent_block_index(
txo_id_hex: &str,
spent_block_index: u64,
conn: &Conn
) -> Result<(), WalletDbError>

Update a Txo’s status to spent

fn update_key_image(
txo_id_hex: &str,
key_image: &KeyImage,
spent_block_index: Option<u64>,
conn: &Conn
) -> Result<(), WalletDbError>

fn list(
status: Option<TxoStatus>,
min_received_block_index: Option<u64>,
max_received_block_index: Option<u64>,
offset: Option<u64>,
limit: Option<u64>,
token_id: Option<u64>,
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

fn list_for_account(
account_id_hex: &str,
status: Option<TxoStatus>,
min_received_block_index: Option<u64>,
max_received_block_index: Option<u64>,
offset: Option<u64>,
limit: Option<u64>,
token_id: Option<u64>,
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

Get all Txos associated with a given account.

fn list_for_address(
assigned_subaddress_b58: &str,
status: Option<TxoStatus>,
min_received_block_index: Option<u64>,
max_received_block_index: Option<u64>,
offset: Option<u64>,
limit: Option<u64>,
token_id: Option<u64>,
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

fn list_unspent_or_pending_key_images(
account_id_hex: &str,
token_id: Option<u64>,
conn: &Conn
) -> Result<HashMap<KeyImage, String>, WalletDbError>

Get a map from key images to unspent txos for this account.

fn list_unspent(
account_id_hex: Option<&str>,
assigned_subaddress_b58: Option<&str>,
token_id: Option<u64>,
min_received_block_index: Option<u64>,
max_received_block_index: Option<u64>,
offset: Option<u64>,
limit: Option<u64>,
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

fn list_spent(
account_id_hex: Option<&str>,
assigned_subaddress_b58: Option<&str>,
token_id: Option<u64>,
min_received_block_index: Option<u64>,
max_received_block_index: Option<u64>,
offset: Option<u64>,
limit: Option<u64>,
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

fn list_orphaned(
account_id_hex: Option<&str>,
token_id: Option<u64>,
min_received_block_index: Option<u64>,
max_received_block_index: Option<u64>,
offset: Option<u64>,
limit: Option<u64>,
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

fn list_pending(
account_id_hex: Option<&str>,
assigned_subaddress_b58: Option<&str>,
token_id: Option<u64>,
min_received_block_index: Option<u64>,
max_received_block_index: Option<u64>,
offset: Option<u64>,
limit: Option<u64>,
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

fn list_unverified(
account_id_hex: Option<&str>,
assigned_subaddress_b58: Option<&str>,
token_id: Option<u64>,
min_received_block_index: Option<u64>,
max_received_block_index: Option<u64>,
offset: Option<u64>,
limit: Option<u64>,
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

fn list_spendable(
account_id_hex: Option<&str>,
max_spendable_value: Option<u64>,
assigned_subaddress_b58: Option<&str>,
token_id: u64,
default_token_fee: u64,
conn: &Conn
) -> Result<SpendableTxosResult, WalletDbError>

fn list_created(
account_id_hex: Option<&str>,
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

fn list_secreted(
account_id_hex: Option<&str>,
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

fn get(txo_id_hex: &str, conn: &Conn) -> Result<Txo, WalletDbError>

Get the details for a specific Txo.

Returns:

Txo

fn select_by_public_key(
public_keys: &[&CompressedRistrettoPublic],
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

Get several Txos by Txo public_keys

Returns:

Vec

fn select_by_id(
txo_ids: &[String],
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

Select several Txos by their TxoIds

Returns:

Vec<(Txo)>

fn select_spendable_txos_for_value(
account_id_hex: &str,
target_value: u128,
max_spendable_value: Option<u64>,
token_id: u64,
default_token_fee: u64,
conn: &Conn
) -> Result<Vec<Txo>, WalletDbError>

Select a set of unspent Txos to reach a given value.

Returns:

Vec

fn validate_confirmation(
account_id: &AccountID,
txo_id_hex: &str,
confirmation: &TxOutConfirmationNumber,
conn: &Conn
) -> Result<bool, WalletDbError>

Validate a confirmation number for a Txo

Returns:

Bool - true if verified

fn scrub_account(account_id_hex: &str, conn: &Conn) -> Result<(), WalletDbError>

fn delete_unreferenced(conn: &Conn) -> Result<(), WalletDbError>

Delete txos which are not referenced by any account or transaction.

fn status(&self, conn: &Conn) -> Result<TxoStatus, WalletDbError>

fn membership_proof(
&self,
ledger_db: &LedgerDB
) -> Result<TxOutMembershipProof, WalletDbError>