Struct mc_full_service::json_rpc::v1::models::txo::Txo

pub struct Txo {

    pub object: String,
    pub txo_id_hex: String,
    pub value_pmob: String,
    pub recipient_address_id: Option<String>,
    pub received_block_index: Option<String>,
    pub spent_block_index: Option<String>,
    pub is_spent_recovered: bool,
    pub received_account_id: Option<String>,
    pub minted_account_id: Option<String>,
    pub account_status_map: Map<String, Value>,
    pub target_key: String,
    pub public_key: String,
    pub e_fog_hint: String,
    pub subaddress_index: Option<String>,
    pub assigned_address: Option<String>,
    pub key_image: Option<String>,
    pub confirmation: Option<String>,

}

An Txo in the wallet.

An Txo is associated with one or two accounts, and can be categorized with different statuses and types in relation to those accounts.

Fields
object: String

String representing the object’s type. Objects of the same type share the same value.

txo_id_hex: String

Unique identifier for the Txo. Constructed from the contents of the TxOut in the ledger representation.

value_pmob: String

Available pico MOB for this account at the current account_block_height. If the account is syncing, this value may change.

recipient_address_id: Option<String>

Unique identifier for the recipient associated account. Only available if direction is “sent”.

received_block_index: Option<String>

Block index in which the txo was received by an account.

spent_block_index: Option<String>

Block index in which the txo was spent by an account.

is_spent_recovered: bool

Flag that indicates if the spent_block_index was recovered from the ledger. This value is null if the txo is unspent. If true, some information may not be available on the txo without user input. If true, the confirmation number will be null without user input.

received_account_id: Option<String>

The account_id for the account which has received this TXO. This account has spend authority.

minted_account_id: Option<String>

The account_id for the account which minted this Txo.

account_status_map: Map<String, Value>

A normalized hash mapping account_id to account objects. Keys include “type” and “status”.

txo_type: With respect to this account, the Txo may be “minted” or “received”.

txo_status: With respect to this account, the Txo may be “unspent”, “pending”, “spent”, “secreted” or “orphaned”. For received Txos received as an assigned address, the lifecycle is “unspent” -> “pending” -> “spent”. For outbound, minted Txos, we cannot monitor its received lifecycle status with respect to the minting account, we note its status as “secreted”. If a Txo is received at an address unassigned (likely due to a recovered account or using the account on another client), the Txo is considered “orphaned” until its address is calculated – in this case, there are manual ways to discover the missing assigned address for orphaned Txos or to recover an entire account.

target_key: String

A cryptographic key for this Txo.

public_key: String

The public key for this txo, can be used as an identifier to find the txo in the ledger.

e_fog_hint: String

The encrypted fog hint for this Txo.

subaddress_index: Option<String>

The assigned subaddress index for this Txo with respect to its received account.

assigned_address: Option<String>

The address corresponding to the subaddress index which was assigned as an intended sender for this Txo.

key_image: Option<String>

A fingerprint of the txo derived from your private spend key materials, required to spend a Txo.

confirmation: Option<String>

A confirmation number that the sender of the Txo can provide to verify that they participated in the construction of this Txo.

Implementations

impl Txo

pub fn new(txo: &Txo, status: &TxoStatus) -> Txo