Function mc_full_service::json_rpc::v2::api::wallet::wallet_api_inner

pub async fn wallet_api_inner<T, FPR>(
    service: &WalletService<T, FPR>,
    command: JsonCommandRequest
) -> Result<JsonCommandResponse, JsonRPCError>where
    T: BlockchainConnection + UserTxConnection + 'static,
    FPR: FogPubkeyResolver + Send + Sync + 'static,

The Wallet API inner method, which handles switching on the method enum.

Note that this is structured this way so that the routes can be defined to take explicit Rocket state, and then pass the service to the inner method. This allows us to properly construct state with Mock Connection Objects in tests. This also allows us to version the overall API easily.