Trait mc_full_service::service::account::AccountService

pub trait AccountService {
    fn create_account(
        &self,
        name: Option<String>,
        fog_report_url: String,
        fog_report_id: String,
        fog_authority_spki: String
    ) -> Result<Account, AccountServiceError>;
    fn import_account(
        &self,
        mnemonic_phrase: String,
        key_derivation_version: u8,
        name: Option<String>,
        first_block_index: Option<u64>,
        next_subaddress_index: Option<u64>,
        fog_report_url: String,
        fog_report_id: String,
        fog_authority_spki: String
    ) -> Result<Account, AccountServiceError>;
    fn import_account_from_legacy_root_entropy(
        &self,
        entropy: String,
        name: Option<String>,
        first_block_index: Option<u64>,
        next_subaddress_index: Option<u64>,
        fog_report_url: String,
        fog_report_id: String,
        fog_authority_spki: String
    ) -> Result<Account, AccountServiceError>;
    fn import_view_only_account(
        &self,
        view_private_key: String,
        spend_public_key: String,
        name: Option<String>,
        first_block_index: Option<u64>,
        next_subaddress_index: Option<u64>
    ) -> Result<Account, AccountServiceError>;
    fn resync_account(
        &self,
        account_id: &AccountID
    ) -> Result<(), AccountServiceError>;
    fn get_view_only_account_import_request(
        &self,
        account_id: &AccountID
    ) -> Result<JsonRPCRequest, AccountServiceError>;
    fn list_accounts(
        &self,
        offset: Option<u64>,
        limit: Option<u64>
    ) -> Result<Vec<Account>, AccountServiceError>;
    fn get_account(
        &self,
        account_id: &AccountID
    ) -> Result<Account, AccountServiceError>;
    fn get_next_subaddress_index_for_account(
        &self,
        account_id: &AccountID
    ) -> Result<u64, AccountServiceError>;
    fn update_account_name(
        &self,
        account_id: &AccountID,
        name: String
    ) -> Result<Account, AccountServiceError>;
    fn sync_account(
        &self,
        account_id: &AccountID,
        txo_ids_and_key_images: Vec<(String, String)>,
        next_subaddress_index: u64
    ) -> Result<(), AccountServiceError>;
    fn remove_account(
        &self,
        account_id: &AccountID
    ) -> Result<bool, AccountServiceError>;
}

Trait defining the ways in which the wallet can interact with and manage accounts.

Required Methods

fn create_account(
&self,
name: Option<String>,
fog_report_url: String,
fog_report_id: String,
fog_authority_spki: String
) -> Result<Account, AccountServiceError>

Creates a new account with default values.

fn import_account(
&self,
mnemonic_phrase: String,
key_derivation_version: u8,
name: Option<String>,
first_block_index: Option<u64>,
next_subaddress_index: Option<u64>,
fog_report_url: String,
fog_report_id: String,
fog_authority_spki: String
) -> Result<Account, AccountServiceError>

Import an existing account to the wallet using the mnemonic.

fn import_account_from_legacy_root_entropy(
&self,
entropy: String,
name: Option<String>,
first_block_index: Option<u64>,
next_subaddress_index: Option<u64>,
fog_report_url: String,
fog_report_id: String,
fog_authority_spki: String
) -> Result<Account, AccountServiceError>

Import an existing account to the wallet using the entropy.

fn import_view_only_account(
&self,
view_private_key: String,
spend_public_key: String,
name: Option<String>,
first_block_index: Option<u64>,
next_subaddress_index: Option<u64>
) -> Result<Account, AccountServiceError>

Import an existing account to the wallet using the mnemonic.

fn resync_account(
&self,
account_id: &AccountID
) -> Result<(), AccountServiceError>

fn get_view_only_account_import_request(
&self,
account_id: &AccountID
) -> Result<JsonRPCRequest, AccountServiceError>

fn list_accounts(
&self,
offset: Option<u64>,
limit: Option<u64>
) -> Result<Vec<Account>, AccountServiceError>

List accounts in the wallet.

fn get_account(
&self,
account_id: &AccountID
) -> Result<Account, AccountServiceError>

Get an account in the wallet.

fn get_next_subaddress_index_for_account(
&self,
account_id: &AccountID
) -> Result<u64, AccountServiceError>

fn update_account_name(
&self,
account_id: &AccountID,
name: String
) -> Result<Account, AccountServiceError>

Update the name for an account.

fn sync_account(
&self,
account_id: &AccountID,
txo_ids_and_key_images: Vec<(String, String)>,
next_subaddress_index: u64
) -> Result<(), AccountServiceError>

complete a sync request for a view only account

fn remove_account(
&self,
account_id: &AccountID
) -> Result<bool, AccountServiceError>

Remove an account from the wallet.