Trait mc_full_service::service::address::AddressService

pub trait AddressService {
    fn assign_address_for_account(
        &self,
        account_id: &AccountID,
        metadata: Option<&str>
    ) -> Result<AssignedSubaddress, AddressServiceError>;
    fn get_address(
        &self,
        address_b58: &str
    ) -> Result<AssignedSubaddress, AddressServiceError>;
    fn get_address_for_account(
        &self,
        account_id: &AccountID,
        index: i64
    ) -> Result<AssignedSubaddress, AddressServiceError>;
    fn get_addresses(
        &self,
        account_id: Option<String>,
        offset: Option<u64>,
        limit: Option<u64>
    ) -> Result<Vec<AssignedSubaddress>, AddressServiceError>;
    fn verify_address(
        &self,
        public_address: &str
    ) -> Result<PublicAddress, AddressServiceError>;
}

Trait defining the ways in which the wallet can interact with and manage addresses.

Required Methods

fn assign_address_for_account(
&self,
account_id: &AccountID,
metadata: Option<&str>
) -> Result<AssignedSubaddress, AddressServiceError>

Creates a new address with default values.

fn get_address(
&self,
address_b58: &str
) -> Result<AssignedSubaddress, AddressServiceError>

Get an assigned subaddress, if it exists.

fn get_address_for_account(
&self,
account_id: &AccountID,
index: i64
) -> Result<AssignedSubaddress, AddressServiceError>

fn get_addresses(
&self,
account_id: Option<String>,
offset: Option<u64>,
limit: Option<u64>
) -> Result<Vec<AssignedSubaddress>, AddressServiceError>

Gets all the addresses for an optionally given account.

fn verify_address(
&self,
public_address: &str
) -> Result<PublicAddress, AddressServiceError>

Verifies whether an address can be decoded from b58.