Trait mc_full_service::service::confirmation_number::ConfirmationService

pub trait ConfirmationService {
    fn get_confirmations(
        &self,
        transaction_log_id: &str
    ) -> Result<Vec<Confirmation>, ConfirmationServiceError>;
    fn validate_confirmation(
        &self,
        account_id: &AccountID,
        txo_id: &TxoID,
        confirmation_hex: &str
    ) -> Result<bool, ConfirmationServiceError>;
}

Trait defining the ways in which the wallet can interact with and manage tonfirmation numbers.

Required Methods

fn get_confirmations(
&self,
transaction_log_id: &str
) -> Result<Vec<Confirmation>, ConfirmationServiceError>

Get the confirmations from the outputs in a transaction log.

fn validate_confirmation(
&self,
account_id: &AccountID,
txo_id: &TxoID,
confirmation_hex: &str
) -> Result<bool, ConfirmationServiceError>

Validate the confirmation number with a given Txo.