Module mc_full_service::service::gift_code

Service for managing gift codes.

Gift codes are onetime accounts that contain a single Txo. They provide a means to send MOB in a way that can be “claimed,” for example, by pasting a QR code for a gift code into a group chat, and the first person to consume the gift code claims the MOB.

Structs

DecodedGiftCode

EncodedGiftCode

Enums

GiftCodeServiceError

GiftCodeStatus

Possible states for a Gift Code in relation to accounts in this wallet.

Traits

GiftCodeService

Trait defining the ways in which the wallet can interact with and manage gift codes.

Functions

decode_transfer_payload

Decode the gift code from b58 to its component parts.