Trait mc_full_service::service::gift_code::GiftCodeService

pub trait GiftCodeService {
    fn build_gift_code(
        &self,
        from_account_id: &AccountID,
        value: u64,
        name: Option<String>,
        input_txo_ids: Option<&Vec<String>>,
        fee: Option<u64>,
        tombstone_block: Option<u64>,
        max_spendable_value: Option<u64>
    ) -> Result<(TxProposal, EncodedGiftCode), GiftCodeServiceError>;
    fn submit_gift_code(
        &self,
        from_account_id: &AccountID,
        gift_code_b58: &EncodedGiftCode,
        tx_proposal: &TxProposal
    ) -> Result<DecodedGiftCode, GiftCodeServiceError>;
    fn get_gift_code(
        &self,
        gift_code_b58: &EncodedGiftCode
    ) -> Result<DecodedGiftCode, GiftCodeServiceError>;
    fn list_gift_codes(
        &self,
        offset: Option<u64>,
        limit: Option<u64>
    ) -> Result<Vec<DecodedGiftCode>, GiftCodeServiceError>;
    fn check_gift_code_status(
        &self,
        gift_code_b58: &EncodedGiftCode
    ) -> Result<(GiftCodeStatus, Option<i64>, String), GiftCodeServiceError>;
    fn claim_gift_code(
        &self,
        gift_code_b58: &EncodedGiftCode,
        account_id: &AccountID,
        public_address_b58: Option<String>
    ) -> Result<Tx, GiftCodeServiceError>;
    fn remove_gift_code(
        &self,
        gift_code_b58: &EncodedGiftCode
    ) -> Result<bool, GiftCodeServiceError>;
}

Trait defining the ways in which the wallet can interact with and manage gift codes.

Required Methods

fn build_gift_code(
&self,
from_account_id: &AccountID,
value: u64,
name: Option<String>,
input_txo_ids: Option<&Vec<String>>,
fee: Option<u64>,
tombstone_block: Option<u64>,
max_spendable_value: Option<u64>
) -> Result<(TxProposal, EncodedGiftCode), GiftCodeServiceError>

Builds a new gift code.

Building a gift code requires the following steps:

Create a new account to receive the funds

Send a transaction to the new account

Wait for the transaction to land

Package the required information into a b58-encoded string

Returns:

JsonSubmitResponse from submitting the gift code transaction to the network

Entropy of the gift code account, hex encoded

fn submit_gift_code(
&self,
from_account_id: &AccountID,
gift_code_b58: &EncodedGiftCode,
tx_proposal: &TxProposal
) -> Result<DecodedGiftCode, GiftCodeServiceError>

fn get_gift_code(
&self,
gift_code_b58: &EncodedGiftCode
) -> Result<DecodedGiftCode, GiftCodeServiceError>

Get the details for a specific gift code.

fn list_gift_codes(
&self,
offset: Option<u64>,
limit: Option<u64>
) -> Result<Vec<DecodedGiftCode>, GiftCodeServiceError>

List all gift codes in the wallet.

fn check_gift_code_status(
&self,
gift_code_b58: &EncodedGiftCode
) -> Result<(GiftCodeStatus, Option<i64>, String), GiftCodeServiceError>

Check the status of a gift code currently in your wallet. If the gift code is not yet in the wallet, add it.

fn claim_gift_code(
&self,
gift_code_b58: &EncodedGiftCode,
account_id: &AccountID,
public_address_b58: Option<String>
) -> Result<Tx, GiftCodeServiceError>

Execute a transaction from the gift code account to drain the account to the destination specified by the account_id_hex and public_address_b58. If no public_address_b58 is provided, then a new AssignedSubaddress will be created to receive the funds.

fn remove_gift_code(
&self,
gift_code_b58: &EncodedGiftCode
) -> Result<bool, GiftCodeServiceError>