Trait mc_full_service::service::ledger::LedgerService

pub trait LedgerService {

    fn get_network_block_height(&self) -> Result<u64, LedgerServiceError>;
    fn get_transaction_object(
        &self,
        transaction_id_hex: &str
    ) -> Result<Tx, LedgerServiceError>;
    fn get_txo_object(
        &self,
        txo_id_hex: &str
    ) -> Result<TxOut, LedgerServiceError>;
    fn get_block_object(
        &self,
        block_index: u64
    ) -> Result<(Block, BlockContents), LedgerServiceError>;
    fn get_block_objects(
        &self,
        first_block_index: u64,
        limit: usize
    ) -> Result<Vec<(Block, BlockContents)>, LedgerServiceError>;
    fn get_recent_block_objects(
        &self,
        limit: usize
    ) -> Result<Vec<(Block, BlockContents)>, LedgerServiceError>;
    fn contains_key_image(
        &self,
        key_image: &KeyImage
    ) -> Result<bool, LedgerServiceError>;
    fn get_latest_block_info(&self) -> Result<BlockInfo, LedgerServiceError>;
    fn get_network_fees(&self) -> Result<FeeMap, LedgerServiceError>;
    fn get_network_block_version(
        &self
    ) -> Result<BlockVersion, LedgerServiceError>;
    fn get_tx_out_proof_of_memberships(
        &self,
        indices: &[u64]
    ) -> Result<Vec<TxOutMembershipProof>, LedgerServiceError>;
    fn get_indices_from_txo_public_keys(
        &self,
        public_keys: &[CompressedRistrettoPublic]
    ) -> Result<Vec<u64>, LedgerServiceError>;
    fn sample_mixins(
        &self,
        num_mixins: usize,
        excluded_indices: &[u64]
    ) -> Result<(Vec<TxOut>, Vec<TxOutMembershipProof>), LedgerServiceError>;
    fn get_block_index_from_txo_public_key(
        &self,
        public_key: &CompressedRistrettoPublic
    ) -> Result<u64, LedgerServiceError>;
    fn search_ledger(
        &self,
        query: &str
    ) -> Result<Vec<LedgerSearchResult>, LedgerServiceError>;

}

Trait defining the ways in which the wallet can interact with and manage ledger objects and interfaces.

Required Methods

fn get_network_block_height(&self) -> Result<u64, LedgerServiceError>

Get the total number of blocks on the ledger.

fn get_transaction_object(
&self,
transaction_id_hex: &str
) -> Result<Tx, LedgerServiceError>

fn get_txo_object(&self, txo_id_hex: &str) -> Result<TxOut, LedgerServiceError>

fn get_block_object(
&self,
block_index: u64
) -> Result<(Block, BlockContents), LedgerServiceError>

fn get_block_objects(
&self,
first_block_index: u64,
limit: usize
) -> Result<Vec<(Block, BlockContents)>, LedgerServiceError>

fn get_recent_block_objects(
&self,
limit: usize
) -> Result<Vec<(Block, BlockContents)>, LedgerServiceError>

fn contains_key_image(
&self,
key_image: &KeyImage
) -> Result<bool, LedgerServiceError>

fn get_latest_block_info(&self) -> Result<BlockInfo, LedgerServiceError>

fn get_network_fees(&self) -> Result<FeeMap, LedgerServiceError>

fn get_network_block_version(&self) -> Result<BlockVersion, LedgerServiceError>

fn get_tx_out_proof_of_memberships(
&self,
indices: &[u64]
) -> Result<Vec<TxOutMembershipProof>, LedgerServiceError>

fn get_indices_from_txo_public_keys(
&self,
public_keys: &[CompressedRistrettoPublic]
) -> Result<Vec<u64>, LedgerServiceError>

fn sample_mixins(
&self,
num_mixins: usize,
excluded_indices: &[u64]
) -> Result<(Vec<TxOut>, Vec<TxOutMembershipProof>), LedgerServiceError>

fn get_block_index_from_txo_public_key(
&self,
public_key: &CompressedRistrettoPublic
) -> Result<u64, LedgerServiceError>

fn search_ledger(
&self,
query: &str
) -> Result<Vec<LedgerSearchResult>, LedgerServiceError>