Enum mc_full_service::service::receipt::ReceiptTransactionStatus

pub enum ReceiptTransactionStatus {
    TransactionSuccess,
    TransactionPending,
    TxosReceivedAtDifferentBlockIndices,
    InvalidConfirmation,
    DuplicateTxos,
    AmountMismatch(String),
    FailedAmountDecryption,
}

Variants

TransactionSuccess

All Txos are in the ledger at the same block index, and the expected value matches the value of the Txos.

TransactionPending

No Txos have landed in the wallet yet.

TxosReceivedAtDifferentBlockIndices

All Txos are in the ledger, at different block indices. This indicates the Txos were spent in different transactions, and the receipt is invalid.

InvalidConfirmation

Invalid confirmation number.

DuplicateTxos

Receipt contains duplicate Txos

AmountMismatch(String)

Receipt Amount does not match the Amount in the Txo: {0}

FailedAmountDecryption

Failed to decrypt the amount for the given Txo