Struct mc_full_service::service::WalletService

pub struct WalletService<T: BlockchainConnection + UserTxConnection + 'static, FPR: FogPubkeyResolver + Send + Sync + 'static> {
    pub wallet_db: Option<WalletDb>,
    pub ledger_db: LedgerDB,
    pub watcher_db: Option<WatcherDB>,
    pub peer_manager: McConnectionManager<T>,
    pub network_state: Arc<RwLock<PollingNetworkState<T>>>,
    pub fog_resolver_factory: Arc<dyn Fn(&[FogUri]) -> Result<FPR, String> + Send + Sync>,
    pub submit_node_offset: Arc<AtomicUsize>,
    pub offline: bool,
    pub logger: Logger,
    /* private fields */
}

Service for interacting with the wallet

Note that some fields need to be pub in order to be used in trait implementations, as Rust sees those as separate modules when they are in separate files.

Fields
wallet_db: Option<WalletDb>

Wallet database handle.

ledger_db: LedgerDB

Ledger database.

watcher_db: Option<WatcherDB>

Watcher database.

peer_manager: McConnectionManager<T>

Peer manager for consensus validators to query for network height.

network_state: Arc<RwLock<PollingNetworkState<T>>>

Representation of the current network state.

fog_resolver_factory: Arc<dyn Fn(&[FogUri]) -> Result<FPR, String> + Send + Sync>

Fog resolver factory to obtain the public key of the ingest enclave from a fog address.

submit_node_offset: Arc<AtomicUsize>

Monotonically increasing counter. This is used for node round-robin selection.

offline: bool

Whether the service should run in offline mode.

logger: Logger

Logger.

Implementations

impl<T: BlockchainConnection + UserTxConnection + 'static, FPR: FogPubkeyResolver + Send + Sync + 'static> WalletService<T, FPR>

pub fn new(
wallet_db: Option<WalletDb>,
ledger_db: LedgerDB,
watcher_db: Option<WatcherDB>,
peer_manager: McConnectionManager<T>,
network_state: Arc<RwLock<PollingNetworkState<T>>>,
fog_resolver_factory: Arc<dyn Fn(&[FogUri]) -> Result<FPR, String> + Send + Sync>,
offline: bool,
logger: Logger
) -> Self

pub fn get_conn(&self) -> Result<Conn, WalletDbError>