Trait mc_full_service::service::transaction_log::TransactionLogService

pub trait TransactionLogService {
    fn list_transaction_logs(
        &self,
        account_id: Option<String>,
        offset: Option<u64>,
        limit: Option<u64>,
        min_block_index: Option<u64>,
        max_block_index: Option<u64>
    ) -> Result<Vec<(TransactionLog, AssociatedTxos, ValueMap)>, WalletServiceError>;
    fn get_transaction_log(
        &self,
        transaction_id_hex: &str
    ) -> Result<(TransactionLog, AssociatedTxos, ValueMap), TransactionLogServiceError>;
}

Trait defining the ways in which the wallet can interact with and manage transaction logs.

Required Methods

fn list_transaction_logs(
&self,
account_id: Option<String>,
offset: Option<u64>,
limit: Option<u64>,
min_block_index: Option<u64>,
max_block_index: Option<u64>
) -> Result<Vec<(TransactionLog, AssociatedTxos, ValueMap)>, WalletServiceError>

List all transactions associated with the given Account ID.

fn get_transaction_log(
&self,
transaction_id_hex: &str
) -> Result<(TransactionLog, AssociatedTxos, ValueMap), TransactionLogServiceError>

Get a specific transaction log.