Trait mc_full_service::service::txo::TxoService

pub trait TxoService {
    fn list_txos(
        &self,
        account_id: Option<String>,
        address: Option<String>,
        status: Option<TxoStatus>,
        token_id: Option<u64>,
        min_received_block_index: Option<u64>,
        max_received_block_index: Option<u64>,
        offset: Option<u64>,
        limit: Option<u64>
    ) -> Result<Vec<(Txo, TxoStatus)>, TxoServiceError>;
    fn get_txo(
        &self,
        txo_id: &TxoID
    ) -> Result<(Txo, TxoStatus), TxoServiceError>;
    fn split_txo(
        &self,
        txo_id: &TxoID,
        output_values: &[String],
        subaddress_index: Option<i64>,
        fee_value: Option<String>,
        fee_token_id: Option<String>,
        tombstone_block: Option<String>
    ) -> Result<TxProposal, TxoServiceError>;
}

Trait defining the ways in which the wallet can interact with and manage Txos.

Required Methods

fn list_txos(
&self,
account_id: Option<String>,
address: Option<String>,
status: Option<TxoStatus>,
token_id: Option<u64>,
min_received_block_index: Option<u64>,
max_received_block_index: Option<u64>,
offset: Option<u64>,
limit: Option<u64>
) -> Result<Vec<(Txo, TxoStatus)>, TxoServiceError>

List the Txos for a given account in the wallet.

fn get_txo(&self, txo_id: &TxoID) -> Result<(Txo, TxoStatus), TxoServiceError>

Get a Txo from the wallet.

fn split_txo(
&self,
txo_id: &TxoID,
output_values: &[String],
subaddress_index: Option<i64>,
fee_value: Option<String>,
fee_token_id: Option<String>,
tombstone_block: Option<String>
) -> Result<TxProposal, TxoServiceError>

Split a Txo